refactor: fix showNPSFeedback logic for multiple feedback types
